move matrix
