+++ packages/perseus-core/src/index.ts
@@ -11,6 +11,7 @@ export type {ErrorKind} from "./error/errors";
 
 // Careful, `version.ts` uses this function so it _must_ be imported above it
 export {addLibraryVersionToPerseusDebug} from "./utils/add-library-version-to-perseus-debug";
+export {default as getMatrixSize} from "./utils/get-matrix-size";
 
 export {libVersion} from "./version";
