+++ packages/perseus-score/src/error-codes.ts
@@ -8,6 +8,7 @@ const MULTIPLICATION_SIGN_ERROR = "MULTIPLICATION_SIGN_ERROR";
 const INVALID_SELECTION_ERROR = "INVALID_SELECTION_ERROR";
 const CHOOSE_CORRECT_NUM_ERROR = "CHOOSE_CORRECT_NUM_ERROR";
 const NOT_NONE_ABOVE_ERROR = "NOT_NONE_ABOVE_ERROR";
+const FILL_ALL_CELLS_ERROR = "FILL_ALL_CELLS_ERROR";
 
 const ErrorCodes = {
     MISSING_PERCENT_ERROR,
@@ -20,6 +21,7 @@ const ErrorCodes = {
     INVALID_SELECTION_ERROR,
     CHOOSE_CORRECT_NUM_ERROR,
     NOT_NONE_ABOVE_ERROR,
+    FILL_ALL_CELLS_ERROR,
 };
 
 export default ErrorCodes;
